Safari-609.1.11.1/LayoutTests/js/dom/Promise-simple-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInInit === firstPromise is false
PASS thisInInit === window is true
PASS thisInFulfillCallback === secondPromise is false
PASS thisInFulfillCallback === window is true
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
